+++ CHANGELOG.md
@@ -1,3 +1,236 @@
+## [1.4.1-dev.5](https://github.com/MorpheApp/morphe-patches-library/compare/v1.4.1-dev.4...v1.4.1-dev.5) (2026-06-21)
+
+
+### Bug Fixes
+
+* Include compatibility app versions ([ea02ae5](https://github.com/MorpheApp/morphe-patches-library/commit/ea02ae5b346925f2999f946381484f85ca2e9b0e))
+
+## [1.4.1-dev.4](https://github.com/MorpheApp/morphe-patches-library/compare/v1.4.1-dev.3...v1.4.1-dev.4) (2026-06-14)
+
+
+### Bug Fixes
+
+* Forward `cancel()` to super to enable `OnCancelListener` ([8d10b99](https://github.com/MorpheApp/morphe-patches-library/commit/8d10b991311dcdc88f8bcc239d0be9c3954e76de))
+
+## [1.4.1-dev.3](https://github.com/MorpheApp/morphe-patches-library/compare/v1.4.1-dev.2...v1.4.1-dev.3) (2026-06-14)
+
+
+### Bug Fixes
+
+* Revert to system dim to fix landscape strips ([520dbe6](https://github.com/MorpheApp/morphe-patches-library/commit/520dbe6460c73d5a909c6e04fba804721e21e322))
+
+## [1.4.1-dev.2](https://github.com/MorpheApp/morphe-patches-library/compare/v1.4.1-dev.1...v1.4.1-dev.2) (2026-06-14)
+
+
+### Bug Fixes
+
+* Fix landscape width and dim in `SheetBottomDialog` ([10d40bb](https://github.com/MorpheApp/morphe-patches-library/commit/10d40bbc1595a7a7fe5f7b5c73c56d3bcfd28790))
+
+## [1.4.1-dev.1](https://github.com/MorpheApp/morphe-patches-library/compare/v1.4.0...v1.4.1-dev.1) (2026-06-13)
+
+
+### Bug Fixes
+
+* Fix dim flash when closing nested `SheetBottomDialog` ([68637ef](https://github.com/MorpheApp/morphe-patches-library/commit/68637efc3ab797482f7c7569eef80d2b36eaf59b))
+
+# [1.4.0](https://github.com/MorpheApp/morphe-patches-library/compare/v1.3.4...v1.4.0) (2026-06-12)
+
+
+### Bug Fixes
+
+* `Dim` metrics values do not update when changing foldable device screen ([#28](https://github.com/MorpheApp/morphe-patches-library/issues/28)) ([0d254c1](https://github.com/MorpheApp/morphe-patches-library/commit/0d254c1a21dd2aa8cda3ecb82337754867aa0288))
+* Cleanup logging ([63c5ce5](https://github.com/MorpheApp/morphe-patches-library/commit/63c5ce570154618827ab2e004741104882cb4801))
+* Fix refactoring ([cb0cdc3](https://github.com/MorpheApp/morphe-patches-library/commit/cb0cdc34470976783dc6155b4dcd2532541b9f29))
+* Ignore replacing `packageName` field ([9e24bc5](https://github.com/MorpheApp/morphe-patches-library/commit/9e24bc531ff604b95e328ca230ea7044f68456bf))
+* Update to latest patcher ([b1051aa](https://github.com/MorpheApp/morphe-patches-library/commit/b1051aaf3724a79ccc8343f112aa4ad2f1ef987d))
+
+
+### Features
+
+* Add `Fingerprint.matchSingle()`` ([1c7f805](https://github.com/MorpheApp/morphe-patches-library/commit/1c7f8057a154b23ad9ae8cbcad9abe8e1849d1af))
+* Add change package installer patch ([136bb71](https://github.com/MorpheApp/morphe-patches-library/commit/136bb711083aa2d397290db366fe7ea5ae30635e))
+* Add ViewAnimations fade in/out utilities ([01027db](https://github.com/MorpheApp/morphe-patches-library/commit/01027db87738609ddda6fe9b11878de284b99b26))
+
+# [1.4.0-dev.8](https://github.com/MorpheApp/morphe-patches-library/compare/v1.4.0-dev.7...v1.4.0-dev.8) (2026-06-01)
+
+
+### Features
+
+* Add `Fingerprint.matchSingle()`` ([1c7f805](https://github.com/MorpheApp/morphe-patches-library/commit/1c7f8057a154b23ad9ae8cbcad9abe8e1849d1af))
+
+# [1.4.0-dev.7](https://github.com/MorpheApp/morphe-patches-library/compare/v1.4.0-dev.6...v1.4.0-dev.7) (2026-05-31)
+
+
+### Bug Fixes
+
+* Update to latest patcher ([b1051aa](https://github.com/MorpheApp/morphe-patches-library/commit/b1051aaf3724a79ccc8343f112aa4ad2f1ef987d))
+
+# [1.4.0-dev.6](https://github.com/MorpheApp/morphe-patches-library/compare/v1.4.0-dev.5...v1.4.0-dev.6) (2026-05-28)
+
+
+### Features
+
+* Add ViewAnimations fade in/out utilities ([01027db](https://github.com/MorpheApp/morphe-patches-library/commit/01027db87738609ddda6fe9b11878de284b99b26))
+
+# [1.4.0-dev.5](https://github.com/MorpheApp/morphe-patches-library/compare/v1.4.0-dev.4...v1.4.0-dev.5) (2026-05-25)
+
+
+### Bug Fixes
+
+* `Dim` metrics values do not update when changing foldable device screen ([#28](https://github.com/MorpheApp/morphe-patches-library/issues/28)) ([0d254c1](https://github.com/MorpheApp/morphe-patches-library/commit/0d254c1a21dd2aa8cda3ecb82337754867aa0288))
+
+# [1.4.0-dev.4](https://github.com/MorpheApp/morphe-patches-library/compare/v1.4.0-dev.3...v1.4.0-dev.4) (2026-05-20)
+
+
+### Bug Fixes
+
+* Fix refactoring ([cb0cdc3](https://github.com/MorpheApp/morphe-patches-library/commit/cb0cdc34470976783dc6155b4dcd2532541b9f29))
+
+# [1.4.0-dev.3](https://github.com/MorpheApp/morphe-patches-library/compare/v1.4.0-dev.2...v1.4.0-dev.3) (2026-05-19)
+
+
+### Bug Fixes
+
+* Ignore replacing `packageName` field ([9e24bc5](https://github.com/MorpheApp/morphe-patches-library/commit/9e24bc531ff604b95e328ca230ea7044f68456bf))
+
+# [1.4.0-dev.2](https://github.com/MorpheApp/morphe-patches-library/compare/v1.4.0-dev.1...v1.4.0-dev.2) (2026-05-18)
+
+
+### Bug Fixes
+
+* Cleanup logging ([63c5ce5](https://github.com/MorpheApp/morphe-patches-library/commit/63c5ce570154618827ab2e004741104882cb4801))
+
+# [1.4.0-dev.1](https://github.com/MorpheApp/morphe-patches-library/compare/v1.3.4...v1.4.0-dev.1) (2026-05-18)
+
+
+### Features
+
+* Add change package installer patch ([136bb71](https://github.com/MorpheApp/morphe-patches-library/commit/136bb711083aa2d397290db366fe7ea5ae30635e))
+
+## [1.3.4](https://github.com/MorpheApp/morphe-patches-library/compare/v1.3.3...v1.3.4) (2026-05-14)
+
+
+### Bug Fixes
+
+* Change all language preferences to use multiline ([dfe9fa3](https://github.com/MorpheApp/morphe-patches-library/commit/dfe9fa32702de2caeac63f9675e0b558b91fac70))
+
+## [1.3.4-dev.1](https://github.com/MorpheApp/morphe-patches-library/compare/v1.3.3...v1.3.4-dev.1) (2026-05-13)
+
+
+### Bug Fixes
+
+* Change all language preferences to use multiline ([9999289](https://github.com/MorpheApp/morphe-patches-library/commit/9999289b11082b86e137bbe81d5ad33ecc255564))
+
+## [1.3.3](https://github.com/MorpheApp/morphe-patches-library/compare/v1.3.2...v1.3.3) (2026-05-12)
+
+
+### Bug Fixes
+
+* Use correct literal format for float/long returnEarly() ([833b2a9](https://github.com/MorpheApp/morphe-patches-library/commit/833b2a9b69bf19beb90ee1ff2c8d8cc8fdd55924))
+
+## [1.3.3-dev.1](https://github.com/MorpheApp/morphe-patches-library/compare/v1.3.2...v1.3.3-dev.1) (2026-05-12)
+
+
+### Bug Fixes
+
+* Use correct literal format for float/long returnEarly() ([833b2a9](https://github.com/MorpheApp/morphe-patches-library/commit/833b2a9b69bf19beb90ee1ff2c8d8cc8fdd55924))
+
+## [1.3.2](https://github.com/MorpheApp/morphe-patches-library/compare/v1.3.1...v1.3.2) (2026-05-11)
+
+
+### Bug Fixes
+
+* Update to latest patcher ([b9bfb4c](https://github.com/MorpheApp/morphe-patches-library/commit/b9bfb4cb1a94af22b381f1f6d593af48eb5367ec))
+
+## [1.3.2-dev.1](https://github.com/MorpheApp/morphe-patches-library/compare/v1.3.1...v1.3.2-dev.1) (2026-05-11)
+
+
+### Bug Fixes
+
+* Update to latest patcher ([b9bfb4c](https://github.com/MorpheApp/morphe-patches-library/commit/b9bfb4cb1a94af22b381f1f6d593af48eb5367ec))
+
+## [1.3.1](https://github.com/MorpheApp/morphe-patches-library/compare/v1.3.0...v1.3.1) (2026-05-10)
+
+
+### Bug Fixes
+
+* Improve logging ([3064bd4](https://github.com/MorpheApp/morphe-patches-library/commit/3064bd4812fc64cf2482cc8699f0a13c2ffb7a50))
+
+## [1.3.1-dev.1](https://github.com/MorpheApp/morphe-patches-library/compare/v1.3.0...v1.3.1-dev.1) (2026-05-10)
+
+
+### Bug Fixes
+
+* Improve logging ([3064bd4](https://github.com/MorpheApp/morphe-patches-library/commit/3064bd4812fc64cf2482cc8699f0a13c2ffb7a50))
+
+# [1.3.0](https://github.com/MorpheApp/morphe-patches-library/compare/v1.2.0...v1.3.0) (2026-05-05)
+
+
+### Bug Fixes
+
+* Patching can crash on Android 14 and lower ([fdea53d](https://github.com/MorpheApp/morphe-patches-library/commit/fdea53d8fb9c5f49338f0b1f0bad3100c59150fb))
+
+
+### Features
+
+* Add remove link verification patch ([81a31f4](https://github.com/MorpheApp/morphe-patches-library/commit/81a31f4182c7f4bed284d1b5e306a64497469001))
+
+# [1.3.0-dev.2](https://github.com/MorpheApp/morphe-patches-library/compare/v1.3.0-dev.1...v1.3.0-dev.2) (2026-05-05)
+
+
+### Bug Fixes
+
+* Patching can crash on Android 14 and lower ([fdea53d](https://github.com/MorpheApp/morphe-patches-library/commit/fdea53d8fb9c5f49338f0b1f0bad3100c59150fb))
+
+# [1.3.0-dev.1](https://github.com/MorpheApp/morphe-patches-library/compare/v1.2.0...v1.3.0-dev.1) (2026-05-04)
+
+
+### Features
+
+* Add remove link verification patch ([81a31f4](https://github.com/MorpheApp/morphe-patches-library/commit/81a31f4182c7f4bed284d1b5e306a64497469001))
+
+# [1.2.0](https://github.com/MorpheApp/morphe-patches-library/compare/v1.1.1...v1.2.0) (2026-04-28)
+
+
+### Features
+
+* Add common helper methods ([#13](https://github.com/MorpheApp/morphe-patches-library/issues/13)) ([964c660](https://github.com/MorpheApp/morphe-patches-library/commit/964c660c2a71bfc6379280a227a3ae1610e4dc31))
+
+# [1.2.0-dev.1](https://github.com/MorpheApp/morphe-patches-library/compare/v1.1.1...v1.2.0-dev.1) (2026-04-27)
+
+
+### Features
+
+* Add common helper methods ([#13](https://github.com/MorpheApp/morphe-patches-library/issues/13)) ([964c660](https://github.com/MorpheApp/morphe-patches-library/commit/964c660c2a71bfc6379280a227a3ae1610e4dc31))
+
+## [1.1.1](https://github.com/MorpheApp/morphe-patches-library/compare/v1.1.0...v1.1.1) (2026-04-27)
+
+
+### Bug Fixes
+
+* Use new Compatibility API in 'PatchListGenerator' ([dc0870d](https://github.com/MorpheApp/morphe-patches-library/commit/dc0870ddeda5d416b24c091189a55fe5eda54ae1))
+
+## [1.1.1-dev.1](https://github.com/MorpheApp/morphe-patches-library/compare/v1.1.0...v1.1.1-dev.1) (2026-04-26)
+
+
+### Bug Fixes
+
+* Use new Compatibility API in 'PatchListGenerator' ([dc0870d](https://github.com/MorpheApp/morphe-patches-library/commit/dc0870ddeda5d416b24c091189a55fe5eda54ae1))
+
+# [1.1.0](https://github.com/MorpheApp/morphe-patches-library/compare/v1.0.4...v1.1.0) (2026-04-26)
+
+
+### Features
+
+* Add `accentOkButton` option to `CustomDialog` ([c8e35f4](https://github.com/MorpheApp/morphe-patches-library/commit/c8e35f492e57ee5928df67fe69310e5f4851c757))
+
+# [1.1.0-dev.1](https://github.com/MorpheApp/morphe-patches-library/compare/v1.0.4...v1.1.0-dev.1) (2026-04-26)
+
+
+### Features
+
+* Add `accentOkButton` option to `CustomDialog` ([c8e35f4](https://github.com/MorpheApp/morphe-patches-library/commit/c8e35f492e57ee5928df67fe69310e5f4851c757))
+
 ## [1.0.4](https://github.com/MorpheApp/morphe-patches-library/compare/v1.0.3...v1.0.4) (2026-04-19)